docs: Modified and fixed readme to reflect latest version
Aswell as doing some fixes on the readme, I made it a bit nicer and specified the use of sqlalchemy and added a link to my configuration management repository which i used to deploy the app.

I also removed the link to Udacity, as I wanted to keep the repository documentation clean and focused on the technology.